+++ src/main/java/org/com/deputatbot/repos/OkrugNduRepo.java
@@ -5,6 +5,7 @@ import org.com.deputatbot.domain.OkrugNdu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface OkrugNduRepo extends JpaRepository<OkrugNdu, Long> {
+public interface OkrugNduRepo extends JpaRepository<OkrugNdu, Integer> {
     OkrugNdu findByNumber(Integer number);
+
 }
