+++ src/main/java/org/com/deputatbot/repos/OkrugOblRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OkrugOblRepo extends JpaRepository<OkrugObl,Long> {
+public interface OkrugOblRepo extends JpaRepository<OkrugObl,Integer> {
    OkrugObl findByNumber(Integer number);
    List<OkrugObl> findAllByNumber(Integer number);
    OkrugObl findByRegionContaining(String dilnizia);
